Modify NancyEngine to allow hosting handler access to StatucCodeHandlers and to make use of consistent exception response handling.
Note that the FindStatusCodeHandler changes effectively incorporate PR #1688 (wasn't sure how to pull it in and modify). I'm not sure if this is a breaking change or not.
The two new methods have not (yet) been added to INancyEngine to avoid any further breaking changes. Ideally these should be added to the interface contract.